% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{expand_rows}
\alias{expand_rows}
\title{Function \code{expand}}
\usage{
expand_rows(df, n = 2, type = c("each", "times"))
}
\arguments{
\item{df}{data frame}

\item{n}{number of duplicates per row}

\item{type}{character scalar. If \code{"each"}, rows will be duplicated in place.
If \code{"times"}, the data frame itself will be repeated \code{n} times}
}
\description{
Expand the rows of a data frame
Copied and modified from \code{drake::expand()} under GPL-3.
}
\examples{
df <- head(mtcars)
expand_rows(mtcars, n = 2, type = "each")
expand_rows(mtcars, n = 2, type = "times")
}
\seealso{
\code{\link{wildcard}}]
}
