% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{expandrows}
\alias{expandrows}
\title{Function \code{expand}}
\usage{
expandrows(df, n = 2, type = c("each", "times"))
}
\arguments{
\item{df}{data frame}

\item{n}{number of duplicates per row}

\item{type}{character scalar. If \code{'each'},
rows will be duplicated in place.
If \code{'times'}, the data frame itself will be repeated \code{n} times.}
}
\description{
Expand the rows of a data frame
Copied and modified from \code{remakeGenerator::expand()} under GPL>=3:
\url{https://github.com/wlandau/remakeGenerator}
}
\examples{
df <- data.frame(
  ID = c('24601', 'Javert', 'Fantine'),
  fate = c('fulfillment', 'confusion', 'misfortune'))
expandrows(df, n = 2, type = 'each')
expandrows(df, n = 2, type = 'times')
}
\seealso{
\code{\link{wildcard}}]
}
