% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{nofactors}
\alias{nofactors}
\title{Function \code{nofactors}}
\usage{
nofactors(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Turn all the factors of a data frame into characters.
}
\examples{
class(iris$Species)
str(iris)
out <- nofactors(iris)
class(out$Species)
str(out)
}
\seealso{
\code{\link{wildcard}}
}
